pyproject.toml
[project]
authors = [
  {name = "Nikhil Kumar", email = "nikumar1206@gmail.com"},
  {name = "Alberto Calderari", email = "actwitter19@gmail.com"},
]

dependencies = [
  "aioboto3>=13.0.0",
  "aiobotocore>=2.13.0",
  "pydantic>=2.7.0",
  "types-aioboto3[dynamodb]>=15.5.0",
]

description = "Async DynamoDB ORM with Pydantic"
dynamic = ["version"]
name = "aiodynamodb"
readme = "README.md"
requires-python = ">=3.12"

[project.optional-dependencies]
testing = [
  "aiomoto>=0.3.0",
  "moto[dynamodb]>=5.1.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[tool.hatch.version]
source = "vcs"

[tool.ruff]
line-length = 120
preview = true
target-version = "py312"

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
command_line = "-m pytest"
source = ["src/aiodynamodb"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
check_untyped_defs = true
python_version = "3.12"
warn_return_any = false
warn_unused_ignores = true

[[tool.mypy.overrides]]
# perf-dev optional dependencies — only present when that group is installed
ignore_missing_imports = true
module = ["matplotlib", "matplotlib.*", "numpy", "pyinstrument"]

[dependency-groups]
# dependency groups are for local development only
dev = [
  "boto3-stubs[dynamodb]>=1.38.0",
  "coverage[toml]>=7.10.6",
  "mypy>=1.19",
  "pytest-cov",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "ruff>=0.14.14",
]
docs = ["zensical==0.0.30"]
perf-dev = [
  "aiomoto>=0.3.0",
  "matplotlib>=3.10.8",
  "numpy>=2.4.3",
  "pyinstrument>=5.1.2",
]
